Move header and allow children to be passed.
